[IMPROVEMENT] Complete redesign of .vimrc
Vim configuration is now modular like never before.
Plugins are configured in an external file so that no more modification
to .vimrc is needed to install a new one.
Plugin directory and Vim data directory can be configured via
environment variables.
Customizations to the base config can be added using a
'vimrc.conf.before' or 'vimrc.conf.after'. This allows to keep the base
configuration in 'vimrc.conf' more stable and allows per-usertweaking.
Added support for local customizations via the dynamic loading of a
'.vimrc.custom' file that will be loaded if found in the directory from
where Vim is started (allows per-project customizations)